requirements.txt
# requirements.txt

# --- Core LangChain ---
langchain
langchain-community
langchain-experimental

# --- LLM & Integrations ---
langchain-groq
langchain-huggingface
langchain-chroma
flashrank

# --- Embeddings ---
sentence-transformers

# --- Data Loaders ---
pypdf
GitPython
pandas

# --- Config & App ---
python-dotenv
pydantic-settings
fastapi
uvicorn

# -- Evaluation --
ragas
deepeval